requirements.txt
Babel==2.7.0
bokeh==1.4.0
Click==7.0
Flask==1.1.1
Flask-Babel==0.12.2
flask-optimize==0.2.9.8
Flask-SQLAlchemy==2.4.1
Flask-Table==0.5.0
htmlmin==0.1.12
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
numpy==1.17.4
packaging==19.2
pandas==1.0.1
Pillow==6.2.1
pyparsing==2.4.5
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
six==1.13.0
SQLAlchemy==1.3.13
tornado==6.0.3
Werkzeug==0.16.0
<<<<<<< HEAD
=======


>>>>>>> 03a11e187f628fbb0ba5cfc15443cbf0eb7cab85
